+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## _(in-progress)_
 * New: Add icon to 'Date' column of post listing to indicate posts that were or will be silently published
+* New: Allow the "Silent Publish?" value to be set and seen in the quick edit pane
+    * New: Add `add_to_quick_edit()`
+    * New: Add `admin_enqueue_scripts()`
 * New: Add TODO.md and move existing TODO list from top of main plugin file into it (and add more items to the list)
 * Change: Show message instead of disabled checkbox when editing silently published post and show nothing for published posts that weren't silently published
 * Change: Enable plugin functionality for all public post types by default